# packages/contracts/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/v1/auth/github": {
      "get": {
        "operationId": "AuthController_signIn",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Iniciar sesión con GitHub",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Cerrar sesión",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "operationId": "AuthController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrentUserDto"
                }
              }
            }
          }
        },
        "summary": "Quién soy",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/workspaces": {
      "get": {
        "operationId": "WorkspacesController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WorkspaceDto"
                  }
                }
              }
            }
          }
        },
        "summary": "Workspaces a los que pertenezco",
        "tags": [
          "workspaces"
        ]
      }
    },
    "/api/v1/workspaces/{id}": {
      "patch": {
        "operationId": "WorkspacesController_rename",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RenameWorkspaceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceDto"
                }
              }
            }
          }
        },
        "summary": "Renombrar un workspace propio",
        "tags": [
          "workspaces"
        ]
      }
    },
    "/api/v1/workspaces/{id}/members": {
      "get": {
        "operationId": "WorkspacesController_members",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MemberDto"
                  }
                }
              }
            }
          }
        },
        "summary": "Miembros del workspace",
        "tags": [
          "workspaces"
        ]
      }
    },
    "/api/v1/workspaces/{id}/members/{userId}": {
      "delete": {
        "operationId": "WorkspacesController_removeMember",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "summary": "Expulsar a un miembro",
        "tags": [
          "workspaces"
        ]
      }
    },
    "/api/v1/workspaces/{id}/leave": {
      "post": {
        "operationId": "WorkspacesController_leave",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "summary": "Abandonar un workspace ajeno",
        "tags": [
          "workspaces"
        ]
      }
    },
    "/api/v1/workspaces/{id}/invitations": {
      "get": {
        "operationId": "WorkspacesController_invitations",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InvitationDto"
                  }
                }
              }
            }
          }
        },
        "summary": "Invitaciones del workspace",
        "tags": [
          "workspaces"
        ]
      },
      "post": {
        "description": "La respuesta es idéntica exista o no una cuenta con ese email: invitar no sirve para averiguar quién usa la plataforma.",
        "operationId": "WorkspacesController_invite",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InviteDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvitationDto"
                }
              }
            }
          }
        },
        "summary": "Invitar por email",
        "tags": [
          "workspaces"
        ]
      }
    },
    "/api/v1/workspaces/invitations/{invitationId}": {
      "delete": {
        "operationId": "WorkspacesController_revoke",
        "parameters": [
          {
            "name": "invitationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "summary": "Revocar una invitación",
        "tags": [
          "workspaces"
        ]
      }
    },
    "/api/v1/workspaces/{workspaceId}/apps": {
      "post": {
        "description": "Nace con su documento de visión listo. Lo que crea un invitado en un workspace ajeno queda en WORKSPACE_WRITE.",
        "operationId": "AppsController_create",
        "parameters": [
          {
            "name": "workspaceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAppDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppSummaryDto"
                }
              }
            }
          }
        },
        "summary": "Crear una app",
        "tags": [
          "apps"
        ]
      },
      "get": {
        "operationId": "AppsController_list",
        "parameters": [
          {
            "name": "workspaceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AppSummaryDto"
                  }
                }
              }
            }
          }
        },
        "summary": "Apps visibles del workspace",
        "tags": [
          "apps"
        ]
      }
    },
    "/api/v1/apps/{id}": {
      "get": {
        "operationId": "AppsController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppSummaryDto"
                }
              }
            }
          }
        },
        "summary": "Ficha de una app",
        "tags": [
          "apps"
        ]
      },
      "patch": {
        "operationId": "AppsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAppDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppSummaryDto"
                }
              }
            }
          }
        },
        "summary": "Editar metadatos",
        "tags": [
          "apps"
        ]
      },
      "delete": {
        "operationId": "AppsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "summary": "Eliminar la app y todo su historial",
        "tags": [
          "apps"
        ]
      }
    },
    "/api/v1/apps/{id}/access-level": {
      "patch": {
        "operationId": "AppsController_changeAccessLevel",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeAccessLevelDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppSummaryDto"
                }
              }
            }
          }
        },
        "summary": "Cambiar el nivel de acceso",
        "tags": [
          "apps"
        ]
      }
    },
    "/api/v1/apps/{id}/archive": {
      "post": {
        "operationId": "AppsController_archive",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppSummaryDto"
                }
              }
            }
          }
        },
        "summary": "Archivar",
        "tags": [
          "apps"
        ]
      }
    },
    "/api/v1/apps/{id}/unarchive": {
      "post": {
        "operationId": "AppsController_unarchive",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppSummaryDto"
                }
              }
            }
          }
        },
        "summary": "Desarchivar",
        "tags": [
          "apps"
        ]
      }
    },
    "/api/v1/apps/{id}/transfer-precursor": {
      "post": {
        "operationId": "AppsController_transfer",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransferPrecursorDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppSummaryDto"
                }
              }
            }
          }
        },
        "summary": "Transferir el rol de precursor",
        "tags": [
          "apps"
        ]
      }
    },
    "/health/live": {
      "get": {
        "operationId": "HealthController_live",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LivenessDto"
                }
              }
            }
          }
        },
        "summary": "Liveness: el proceso responde",
        "tags": [
          "health"
        ]
      }
    },
    "/health/ready": {
      "get": {
        "operationId": "HealthController_ready",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadinessReportDto"
                }
              }
            }
          },
          "503": {
            "description": "Alguna dependencia falla",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadinessReportDto"
                }
              }
            }
          }
        },
        "summary": "Readiness: las dependencias responden",
        "tags": [
          "health"
        ]
      }
    }
  },
  "info": {
    "title": "App Foundry API",
    "description": "Un espacio para pensar, definir y traquear ideas de aplicaciones",
    "version": "1.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "schemas": {
      "CurrentUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "handle": {
            "type": "string",
            "description": "Nombre de usuario de GitHub, usado en las menciones"
          },
          "email": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "avatarUrl": {
            "type": "string",
            "nullable": true
          },
          "platformRole": {
            "type": "string",
            "enum": [
              "ADMIN",
              "MEMBER"
            ]
          }
        },
        "required": [
          "id",
          "handle",
          "email",
          "displayName",
          "avatarUrl",
          "platformRole"
        ]
      },
      "WorkspaceDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "description": "Rol de quien consulta en este workspace",
            "enum": [
              "OWNER",
              "MEMBER"
            ]
          },
          "isPersonal": {
            "type": "boolean",
            "description": "Si es el workspace personal de su dueño"
          }
        },
        "required": [
          "id",
          "name",
          "slug",
          "role",
          "isPersonal"
        ]
      },
      "RenameWorkspaceDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 80
          }
        },
        "required": [
          "name"
        ]
      },
      "MemberDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "handle": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "avatarUrl": {
            "type": "string",
            "nullable": true
          },
          "role": {
            "type": "string",
            "enum": [
              "OWNER",
              "MEMBER"
            ]
          }
        },
        "required": [
          "userId",
          "handle",
          "displayName",
          "avatarUrl",
          "role"
        ]
      },
      "InvitationDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "ACCEPTED",
              "REVOKED",
              "EXPIRED"
            ]
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "email",
          "status",
          "expiresAt",
          "createdAt"
        ]
      },
      "InviteDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "description": "Email de la persona a invitar"
          }
        },
        "required": [
          "email"
        ]
      },
      "CreateAppDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 80
          },
          "shortDescription": {
            "type": "string",
            "maxLength": 160
          },
          "accessLevel": {
            "type": "string",
            "enum": [
              "PRIVATE",
              "WORKSPACE_READ",
              "WORKSPACE_WRITE"
            ],
            "description": "Solo lo elige el dueño del workspace. Lo que crea un invitado queda en WORKSPACE_WRITE."
          }
        },
        "required": [
          "name"
        ]
      },
      "AppIconDto": {
        "type": "object",
        "properties": {
          "emoji": {
            "type": "string",
            "description": "Emoji de la selección curada"
          },
          "color": {
            "type": "string",
            "description": "Color de fondo de la paleta"
          }
        },
        "required": [
          "emoji",
          "color"
        ]
      },
      "AppSummaryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "slug": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "shortDescription": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": [
              "IDEA",
              "DEFINING",
              "IN_DEVELOPMENT",
              "PUBLISHED",
              "PAUSED",
              "ARCHIVED"
            ]
          },
          "accessLevel": {
            "type": "string",
            "enum": [
              "PRIVATE",
              "WORKSPACE_READ",
              "WORKSPACE_WRITE"
            ]
          },
          "icon": {
            "$ref": "#/components/schemas/AppIconDto"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "repoUrl": {
            "type": "string",
            "nullable": true
          },
          "precursorHandle": {
            "type": "string",
            "description": "Handle de GitHub de quien la creó"
          },
          "isPrecursor": {
            "type": "boolean",
            "description": "Si quien consulta es su precursor"
          },
          "canEdit": {
            "type": "boolean",
            "description": "Si quien consulta puede editarla"
          },
          "isArchived": {
            "type": "boolean"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "slug",
          "name",
          "shortDescription",
          "status",
          "accessLevel",
          "icon",
          "tags",
          "repoUrl",
          "precursorHandle",
          "isPrecursor",
          "canEdit",
          "isArchived",
          "updatedAt"
        ]
      },
      "UpdateAppDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 80
          },
          "shortDescription": {
            "type": "string",
            "maxLength": 160,
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": [
              "IDEA",
              "DEFINING",
              "IN_DEVELOPMENT",
              "PUBLISHED",
              "PAUSED",
              "ARCHIVED"
            ]
          },
          "tags": {
            "maxItems": 8,
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "repoUrl": {
            "type": "string",
            "description": "Enlace al repositorio. Informativo: no sincroniza nada todavía (RF-417)."
          },
          "iconEmoji": {
            "type": "string",
            "enum": [
              "💡",
              "✨",
              "🌱",
              "🔮",
              "🚀",
              "🧭",
              "🛠️",
              "⚙️",
              "🧰",
              "🔧",
              "📐",
              "🧪",
              "📚",
              "📝",
              "🧠",
              "🔍",
              "🗺️",
              "📊",
              "💬",
              "📣",
              "📨",
              "🤝",
              "🎙️",
              "📡",
              "🏗️",
              "🧱",
              "🗂️",
              "🎛️",
              "🪟",
              "🧩",
              "🌊",
              "🌲",
              "⛰️",
              "🌙",
              "☀️",
              "⏳",
              "🎯",
              "🎲",
              "🎨",
              "🎵",
              "🏔️",
              "🔥"
            ]
          },
          "iconColor": {
            "type": "string",
            "enum": [
              "amber",
              "rose",
              "violet",
              "indigo",
              "sky",
              "teal",
              "emerald",
              "lime",
              "orange",
              "slate"
            ]
          }
        }
      },
      "ChangeAccessLevelDto": {
        "type": "object",
        "properties": {
          "accessLevel": {
            "type": "string",
            "enum": [
              "PRIVATE",
              "WORKSPACE_READ",
              "WORKSPACE_WRITE"
            ]
          }
        },
        "required": [
          "accessLevel"
        ]
      },
      "TransferPrecursorDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "format": "uuid",
            "description": "Miembro del workspace que pasa a ser precursor"
          }
        },
        "required": [
          "userId"
        ]
      },
      "DependencyCheckDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "up",
              "down"
            ],
            "description": "Estado de la dependencia"
          },
          "latencyMs": {
            "type": "number",
            "description": "Tiempo de respuesta en milisegundos",
            "example": 12
          },
          "error": {
            "type": "string",
            "description": "Motivo del fallo, si lo hubo"
          }
        },
        "required": [
          "status",
          "latencyMs"
        ]
      },
      "LivenessDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "ok"
            ]
          }
        },
        "required": [
          "status"
        ]
      },
      "ReadinessReportDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "ok",
              "degraded"
            ]
          },
          "checks": {
            "type": "object",
            "additionalProperties": {
              "$ref": "#/components/schemas/DependencyCheckDto"
            },
            "description": "Una entrada por dependencia comprobada"
          }
        },
        "required": [
          "status",
          "checks"
        ]
      }
    }
  }
}